+++ projects/ngx-location-picker/src/lib/components/ngx-location-picker.component.ts
@@ -33,10 +33,11 @@ export class NgxLocationPickerComponent implements OnInit, OnDestroy, ControlVal
   @Input() baseUrl;
   /* The default zoom level on map load. */
   @Input() defaultZoom = 14;
-  /* Change zoom level on location select to onSelectZoom value */
-  @Input() changeZoomLevelToSelectZoom = true;
-  /* The zoom level when a location is selected. */
-  @Input() onSelectZoom = 16;
+  /**
+   *  The zoom level when a location is selected.
+   *  If null the zoomlevel won't change after location selection.
+   */
+  @Input() onSelectZoom? = 16;
   /* The zoom level will change after location selected (to fit selected geometry). */
   @Input() changeZoomLevelOnLocationSelect = false;
   /* The initial map center on load. */
@@ -714,7 +715,7 @@ export class NgxLocationPickerComponent implements OnInit, OnDestroy, ControlVal
    * Sets leaflet map view to selected coords. With preferred zoomlevel.
    */
   private setView(coords: number[]) {
-    if (this.changeZoomLevelToSelectZoom) {
+    if (this.onSelectZoom) {
       this.leafletMap.setView(coords, this.onSelectZoom);
     } else {
       this.leafletMap.setView(coords, this.leafletMap.map.getZoom());